un-rename GUI and undo Makefile changes, will be in a separate PR
