+++ project/DOS/Makefile
@@ -1,10 +1,10 @@
 bin = MicroWeb.exe
 SRC_PATH = ..\..\src
 OBJDIR=obj
-objects = MicroWeb.obj App.obj Parser.obj Renderer.obj Tags.obj Platform.obj CGA.obj EGA.obj Hercules.obj TextMode.obj Olivetti.obj Font.obj GUI.obj DOSInput.obj DOSNet.obj Page.obj 
+objects = MicroWeb.obj App.obj Parser.obj Renderer.obj Tags.obj Platform.obj CGA.obj EGA.obj Hercules.obj Olivetti.obj TextMode.obj Font.obj Interface.obj DOSInput.obj DOSNet.obj Page.obj 
 memory_model = -ml
 CC = wpp
-CFLAGS = -zq -0 -bt=DOS $(memory_model)
+CFLAGS = -zq -0 -ot -bt=DOS -w2 $(memory_model)
 LD = wlink
 
 # begin mTCP stuff
@@ -13,7 +13,7 @@ tcp_c_dir = ..\..\lib\mTCP\TCPLIB\
 
 tcpobjs = packet.obj arp.obj eth.obj ip.obj tcp.obj tcpsockm.obj udp.obj utils.obj dns.obj timer.obj ipasm.obj trace.obj
 
-tcp_compile_options = -0 $(memory_model) -DCFG_H="tcp.cfg" -s -ei -zp2 -zpw -we
+tcp_compile_options = -0 $(memory_model) -DCFG_H="tcp.cfg" -oh -ok -ot -s -oa -ei -zp2 -zpw -we -ob -ol+ -oi+
 tcp_compile_options += -i=$(tcp_h_dir)
 
 .cpp : $(tcp_c_dir)
@@ -28,7 +28,7 @@ tcp_compile_options += -i=$(tcp_h_dir)
 # end mTCP stuff
 
 $(bin): $(objects) $(tcpobjs)
-    $(LD) system dos name $@ file { *.obj }
+	$(LD) system dos name $@ file { $(objects) $(tcpobjs) }
 
 
 MicroWeb.obj: $(SRC_PATH)\MicroWeb.cpp
@@ -55,7 +55,7 @@ Platform.obj: $(SRC_PATH)\DOS\Platform.cpp
 Font.obj: $(SRC_PATH)\Font.cpp
 	$(CC) -fo=$@ $(CFLAGS) $<
 
-GUI.obj: $(SRC_PATH)\GUI.cpp
+Interface.obj: $(SRC_PATH)\Interface.cpp
 	$(CC) -fo=$@ $(CFLAGS) $<
 
 CGA.obj: $(SRC_PATH)\DOS\CGA.cpp
@@ -64,13 +64,13 @@ CGA.obj: $(SRC_PATH)\DOS\CGA.cpp
 Hercules.obj: $(SRC_PATH)\DOS\Hercules.cpp
 	$(CC) -fo=$@ $(CFLAGS) $<
 
+Olivetti.obj: $(SRC_PATH)\DOS\Olivetti.cpp
+	$(CC) -fo=$@ $(CFLAGS) $<
+
 EGA.obj: $(SRC_PATH)\DOS\EGA.cpp
 	$(CC) -fo=$@ $(CFLAGS) $<
 
 TextMode.obj: $(SRC_PATH)\DOS\TextMode.cpp
-	$(CC) -fo=$@ $(CFLAGS) $<
-
-Olivetti.obj: $(SRC_PATH)\DOS\Olivetti.cpp
 	$(CC) -fo=$@ $(CFLAGS) $<
 
 DOSInput.obj: $(SRC_PATH)\DOS\DOSInput.cpp
