+++ src/DOS/Hercules.cpp
@@ -21,7 +21,7 @@
 #include "../Image.h"
 #include "Hercules.h"
 #include "DefData.h"
-#include "../GUI.h"
+#include "../Interface.h"
 
 #define BASE_VRAM_ADDRESS (uint8_t*) MK_FP(0xB000, 0)
